+++ src/controllers/OrdersController.php
@@ -1333,7 +1333,9 @@ class OrdersController extends Controller
         }
 
         $shippingMethod = $order->shippingMethodHandle ? Plugin::getInstance()->getShippingMethods()->getShippingMethodByHandle($order->shippingMethodHandle) : null;
-        $order->shippingMethodName = $shippingMethod->name ?? null;
+        if ($shippingMethod) {
+            $order->shippingMethodName = $shippingMethod->name ?? null;
+        }
 
         $order->clearNotices();