+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Fixed missing validation for Inventory Location handles. ([#3511](https://github.com/craftcms/commerce/issues/3511))
+- Fixed a bug where an Order’s shipping method name could be cleared when saving a completed order with a shipping method provided by a plugin. ([#3519](https://github.com/craftcms/commerce/issues/3519))
+- Fixed a PHP error that could occur when initializing Craft Commerce.
 
 ## 5.0.7 - 2024-05-22